% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sep_helpers.R
\name{calc_summ_stats_as}
\alias{calc_summ_stats_as}
\title{Calculate gWbar and sigmahat for use when bootstrapping
(Additively Seperable case)}
\usage{
calc_summ_stats_as(W, g)
}
\arguments{
\item{W}{Data Matrix}

\item{g}{the function g(W)}
}
\value{

}
\description{
Calculate gWbar and sigmahat for use when bootstrapping
(Additively Seperable case)
}
