% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sep_helpers.R
\name{gen_Astarinf_obj_as}
\alias{gen_Astarinf_obj_as}
\title{Generate Astarinf object.}
\usage{
gen_Astarinf_obj_as(
  gWbar,
  sigmahat,
  rhatinf,
  sd_gWstarbars,
  sd_sigmastars,
  n,
  kappa,
  f,
  iota_sd
)
}
\arguments{
\item{gWbar}{colmeans of gW}

\item{f}{function f(theta)}

\item{iota_sd}{buffer term (default is 1e-06)}
}
\value{

}
\description{
Parameters are identical for this whole class of functions,
small changes are made in the calculation for different particular
cases, I'm not sure how they differ.
}
