% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sep_helpers.R
\name{mhat_fun_as}
\alias{mhat_fun_as}
\title{mhat calculation}
\usage{
mhat_fun_as(theta, gWbar, sigmahat, f)
}
\arguments{
\item{gWbar}{Column Means of the gW matrix or the mean of g(W) over n}

\item{sigmahat}{Standard Deviation}

\item{f}{The function f(theta) that is a component of m(W,theta)}
}
\value{

}
\description{
mhat calculation
}
