% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momineq.R
\name{momineq_sim}
\alias{momineq_sim}
\title{Implementing SPUR1, SPUR2, and GMS}
\usage{
momineq_sim(
  W,
  k1,
  k2,
  alpha = 0.05,
  alpha2 = 0.045,
  S = S1,
  B = 1000,
  sdS = 250,
  Theta = c(-20, 20),
  theta0s,
  r = NULL,
  length_id = NULL
)
}
\description{
This is the function that runs the inference method introduced by Andrews
and Kwon (2019) but the focus is to replicate the simulations given in the
paper. That is, this function is similar to \code{\link{momineq}} but always
runs SPUR1, SPUR2, and GMS (when r = 0). The following function is for the
lower/upper bound model, and is specialized for the simulation.
}
