% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univ_helpers.R
\name{neg}
\alias{neg}
\title{neg}
\usage{
neg(x)
}
\arguments{
\item{x}{vector of numerics}
}
\value{
vector x with positive values replaced by zeros
}
\description{
neg
}
\examples{
neg(-2:2) == c(-2,-1,0,0,0)
}
