% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univ_helpers.R
\name{phi}
\alias{phi}
\title{phi}
\usage{
phi(xi)
}
\arguments{
\item{xi}{vector of numerics}
}
\value{
vector where any xi>1 is infinity and otherwise is zero
}
\description{
phi
}
\examples{
phi(-2:2) == c(0,0,0,0,Inf)
}
