% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univ_helpers.R
\name{pmax1}
\alias{pmax1}
\title{pmax1}
\usage{
pmax1(x)
}
\arguments{
\item{x}{vector of numerics}
}
\value{
vector x where any value > 1 is replaced by 1
}
\description{
pmax1
}
\examples{
pmax1(-2:2) == c(-2,-1,0,1,1)
}
