% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univ_helpers.R
\name{pos}
\alias{pos}
\title{Positive Elements Only}
\usage{
pos(x)
}
\arguments{
\item{x}{A vector of numerics}
}
\value{
A new vector where the negative values in x are
replaced by zeros
}
\description{
Keeps only the positive elements from the vector and replaces
all negative elements with zeroes.
}
\examples{
pos(-2:2) == c(0,0,0,1,2)
}
\seealso{
\code{\link[=neg]{neg()}}
}
