% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momineq.R
\name{spur}
\alias{spur}
\title{Spurious Precision Calculation}
\usage{
spur(
  W,
  m = NULL,
  additive = FALSE,
  g = NULL,
  f = NULL,
  alpha = 0.05,
  alpha2 = 0.045,
  S = S1,
  B = 1000,
  Theta,
  theta0s,
  rho = NULL
)
}
\arguments{
\item{W}{Matrix of the data (n*k)}

\item{m}{The Function of W and theta, if additively separable leave as NULL}

\item{additive}{Boolean indicating whether the function m
is additively separable.
i.e. it can be represented as m(W,theta) = g(W) + f(theta)}

\item{g}{The function g(W)}

\item{f}{The function f(theta)}

\item{rho}{}
}
\value{

}
\description{
Spurious Precision Calculation
}
