% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spur.R
\name{spur}
\alias{spur}
\title{Spurious Precision Calculation}
\usage{
spur(
  W,
  m = NULL,
  additive = FALSE,
  g = NULL,
  f = NULL,
  Theta,
  theta0s = NULL,
  num_init_val = 1,
  alpha = 0.05,
  alpha2 = 0.045,
  S = S1,
  B = 1000,
  sdS = 250,
  c_tau = 1,
  c_kappa = 1,
  iota_sd = 1e-06,
  iota_q = 1e-06
)
}
\arguments{
\item{W}{Matrix of the data (n by k)}

\item{m}{The Function of W and theta, if additively separable leave as NULL}

\item{additive}{Boolean indicating whether the function m
is additively separable.
i.e. it can be represented as m(W,theta) = g(W) + f(theta)}

\item{g}{The function g(W)}

\item{f}{The function f(theta)}

\item{Theta}{A matrix representing the parameter space of Theta.
The matrix should be 2 by k with the first row representing the lower bounds
of Theta and the second row the upper bounds.}

\item{num_init_val}{Number of initial values to be used in each
bootstrap sample optimization. Choosing values above the default of one
may increase precision, but will also increase calculation time.}

\item{alpha}{Significance level of the SPUR2 test, default is 0.05}

\item{alpha2}{Significance level used for the SPUR1 and GMS tests,
default is 0.045.

alpha = alpha1 + alpha2 so alpha1 is computed by the function given the users
inputs of alpha and alpha2. alpha1 is the significane level used to construct
the CI for r_inf.}

\item{B}{Number of Bootstrap samples used. Default is 1000}

\item{c_tau}{Tau scaling parameter. The default value of Tau is sqrt(log(n)),
changing c_tau scales Tau up or down.}

\item{c_kappa}{Kappa scaling parameter. Identical to c_tau except for kappa.
Kappa's default value is also sqrt(log(n))}

\item{iota_q}{}
}
\value{

}
\description{
Spurious Precision Calculation
}
