% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momineq.R
\name{spur_add_sep}
\alias{spur_add_sep}
\title{Spurious Precision Calculation for Additively Separable Case}
\usage{
spur_add_sep(W, g, f, alpha, alpha2, S, B, Theta, theta0s, rho)
}
\arguments{
\item{...}{All params are passed down from the spur function call.}
}
\value{

}
\description{
Spurious Precision Calculation for Additively Separable Case
}
